services/api_docs/functions/swagger_ui/requirements.txt
boto3==1.28.85
pyyaml==6.0 